==> 1_file_merger.Rmd <==
---
title: "Data_Cleaning"
author: "Tara McAllister Byun"
date: "May 9, 2016"
output: html_document
---

Merging of data (after initial cleaning to eliminate failed raters, null responses, items with <8 trials)
was executed in the file merge_and_summarize_BFS_FINAL.R in C:\Users\tara\Dropbox\R\BFS. This process has not been replicated in the active project folder because it involves looping through embedded subfolders. 

The outputs of the file are BFS2_trials_unsummarized.csv and BFS2_trials_summarized.csv. Both need to be cleaned with 2_clean_data.R.
